Add files via upload
Finalized version of word analyzer. For word count, analyzes given text file, macbeth.txt, and displays word and the corresponding number of occurrences of each word to the user. For database spell check, analyzes given text file, testStates.txt, and compares with dictionary.txt, which are all of the words that have been placed in wordoccurrences schema words table. The program displays the words from test file that are not in the database.